Reuse PCB component bounds helper after layout
