+++ lib/components/primitive-components/Group/Group_doInitialPcbLayoutPack/Group_doInitialPcbLayoutPack.ts
@@ -12,7 +12,7 @@ import Debug from "debug"
 import { applyComponentConstraintClusters } from "./applyComponentConstraintClusters"
 import { applyPackOutput } from "./applyPackOutput"
 import type { NormalComponent } from "lib/components/base-components/NormalComponent"
-import { getPostLayoutBoundsOfPcbComponents } from "lib/utils/get-post-layout-bounds-of-pcb-components"
+import { getBoundsOfPcbComponents } from "lib/utils/get-bounds-of-pcb-components"
 
 const DEFAULT_MIN_GAP = "1mm"
 const debug = Debug("Group_doInitialPcbLayoutPack")
@@ -224,7 +224,9 @@ export const Group_doInitialPcbLayoutPack = (group: Group) => {
 
   applyPackOutput(group, packOutput, clusterMap, initialPackOutput)
   group.calculatePcbGroupBounds(
-    getPostLayoutBoundsOfPcbComponents(group.children),
+    getBoundsOfPcbComponents(group.children, {
+      pcbLayoutPhase: "after_layout",
+    }),
   )
 
   if (packingFailed) return
